requirements.txt
aiohttp
aiosignal
anyio
asgiref
async-timeout
attrs
Babel
beautifulsoup4
blinker
certifi
cffi
charset-normalizer
click
crispy-bootstrap5
cryptography
defusedxml
deprecation
dj-database-url
dj-rest-auth
Django
django-allauth
django-bootstrap3
django-ckeditor
django-cors-headers
django-crispy-forms
django-email-verification
django-environ
django-js-asset
django-rest-framework
drf-yasg
frozenlist
gunicorn
h11
httpcore
httpx
idna
inflection
itsdangerous
jsonschema
jsonschema-specifications
MarkupSafe
multidict
nose
oauthlib
openai
packaging
phonenumbers
phonenumberslite
Pillow
psycopg2
py-moneyed
pycparser
PyJWT
python3-openid
pytz
PyYAML
referencing
requests
requests-oauthlib
rpds-py
sniffio
soupsieve
sqlparse
tqdm
typing_extensions
uritemplate
urllib3
Werkzeug
whitenoise
yarl